virtio_ring: remove a level of indirection
Patchwork-id: 25176
O-Subject: [PATCH RHEL6.0 12/16] virtio_ring: remove a level of indirection
Bugzilla: 593158
RH-Acked-by: Amit Shah <amit.shah@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

We have a single virtqueue_ops implementation,
and it seems unlikely we'll get another one
at this point. So let's remove an unnecessary
level of indirection: it would be very easy to
re-add it if another implementation surfaces.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>

Upstream status
This patch is on -rr tree, intended for 2.6.35.

Test status
The patch was tested in-house on x86_64.

Brew build:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2451100

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>