vhost-net: utilize PUBLISH_USED_IDX feature
Patchwork-id: 25180
O-Subject: [PATCH RHEL6.0 16/16] vhost-net: utilize PUBLISH_USED_IDX feature
Bugzilla: 593158
RH-Acked-by: Amit Shah <amit.shah@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

With PUBLISH_USED_IDX, guest tells us which used entries
it has consumed. This can be used to reduce the number
of interrupts: after we write a used entry, if the guest has not yet
consumed the previous entry, or if the guest has already consumed the
new entry, we do not need to interrupt.
This imporves bandwidth by 30% under some workflows.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>

Upstream status
Posted, acked by DaveM. Will be merged after virtio patch due to
dependency on that.

Test status
The patch was tested in-house on x86_64.

Brew build:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2451100

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>