dasd: fix race between tasklet and dasd_sleep_on
Patchwork-id: 25182
O-Subject: [RHEL6 PATCH 1/1] [s390x] dasd: fix race between tasklet and
	dasd_sleep_on
Bugzilla: 591963
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Description
-----------
A panic might occur when calling the callback function of an cqr.
The various dasd_sleep_on functions use a global wait queue when waiting
for a cqr.  The wait condition checks the status and devlist fields of
the cqr to determine if it is safe to continue.  This evaluation may
return true, although the tasklet has not finished processing of the cqr
and the callback function has not been called yet.
When the callback is finally called, the data in the cqr may already be
invalid.

The sleep_on wait condition needs a safe way to determine if the tasklet
has finished processing.  Use the callback_data field of the cqr to store
a token, which is set by the callback function itself.

Bugzilla
--------
BZ 591963
https://bugzilla.redhat.com/show_bug.cgi?id=591963

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.34
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=1c1e093cbf6d3a7576ba0bd10363362a1c5c74ee

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.