mpt2sas: update to 05.100.00.02
Patchwork-id: 25189
O-Subject: [RHEL6 PATCH] mpt2sas: update to 05.100.00.02
Bugzilla: 591971
RH-Acked-by: Rob Evers <revers@redhat.com>

This is for bz#591971

This patch comes from LSI and as usually also testing and support.
An explanation why we should take the driver so late:
"The reason behind getting mpt2sas up to the latest is that the customer
has a new cluster that they're testing out *now* with the RHEL6 betas.
Of course, it will be a while before this cluster goes live, but, they
have informed us that the mpt2sas stuff is on their critical path to
getting storage built for this big cluster.
150 hard drives behind mptsas fabric (10 ports, 2 per card, each with 15
drives) has been the typical setup.
We're basically trying to get all issues proactively squashed before the
RHEL6 GA goes out and the customer deploys this into production."

The customer is llnl.gov, patch is in scsi-misc tree.
I verified basic working on x86_64, without issues.
Brew build - https://brewweb.devel.redhat.com/taskinfo?taskID=2448533

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>