cifs: guard against hardlinking directories
Patchwork-id: 25190
O-Subject: [RHEL6 PATCH] BZ#591229: cifs: guard against hardlinking directories
Bugzilla: 591229
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

This patch just made 2.6.34 and fixes a reproducible oops at umount. The
problem is that the server can send the same "uniqueid" value to the
client (which we use in cifs as an inode number). The fix is to detect
when this occurs and switch to client-generated inode numbers.

-----------------------------[snip]----------------------------

When we made serverino the default, we trusted that the field sent by the
server in the "uniqueid" field was actually unique. It turns out that it
isn't reliably so.

Samba, in particular, will just put the st_ino in the uniqueid field when
unix extensions are enabled. When a share spans multiple filesystems, it's
quite possible that there will be collisions. This is a server bug, but
when the inodes in question are a directory (as is often the case) and
there is a collision with the root inode of the mount, the result is a
kernel panic on umount.

Fix this by checking explicitly for directory inodes with the same
uniqueid. If that is the case, then we can assume that using server inode
numbers will be a problem and that they should be disabled.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>