cifs: don't update uniqueid in cifs_fattr_to_inode
Patchwork-id: 25206
O-Subject: [RHEL6 PATCH 17/18] BZ#591483: cifs: don't update uniqueid in
	cifs_fattr_to_inode
Bugzilla: 591483
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

We use this value to find an inode within the hash bucket, so we can't
change this without re-hashing the inode. For now, treat this value
as immutable.

Eventually, we should probably use an inode number change on a path
based operation to indicate that the lookup cache is invalid, but that's
a bit more code to deal with.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>