bitmap: update dirty flag when bitmap bits are explicitly set
Patchwork-id: 25393
O-Subject: [Patch RHEL6 15/60] md/bitmap: update dirty flag when bitmap bits are
	explicitly set.
Bugzilla: 583050 586296 586299 588371
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: NeilBrown <neilb@suse.de>

There is a sysfs file which allows bits in the write-intent
bitmap to be explicit set - indicating that the block is thought
to be 'dirty'.
When this happens we should really set recovery_cp backwards
to include the block to reflect this dirtiness.

In particular, a 'resync' process will refuse to start if
recovery_cp is beyond the end of the array, so this is needed
to allow a resync to be triggered.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>