integrate spares into array at earliest opportunity
Patchwork-id: 25368
O-Subject: [Patch RHEL6 20/60] md: integrate spares into array at earliest
	opportunity.
Bugzilla: 583050 586296 586299 588371
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: NeilBrown <neilb@suse.de>

As v1.x metadata can record that a member of the array is
not completely recovered, it make sense to record that a
spare has become a regular member of the array at the earliest
opportunity.
So remove the tests on "recovery_offset > 0" in super_1_sync
as they really aren't needed, and schedule a metadata update
immediately after adding spares to a degraded array.

This means that if a crash happens immediately after a recovery
starts, the new device will be included in the array and recovery will
continue from wherever it was up to.  Previously this didn't happen
unless recovery was at least 1/16 of the way through.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>