deal with merge_bvec_fn in component devices better
Patchwork-id: 25370
O-Subject: [Patch RHEL6 30/60] md: deal with merge_bvec_fn in component devices
	better.
Bugzilla: 583050 586296 586299 588371
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

If a component device has a merge_bvec_fn then as we never call it
we must ensure we never need to.  Currently this is done by setting
max_sector to 1 PAGE, however this does not stop a bio being created
with several sub-page iovecs that would violate the merge_bvec_fn.

So instead set max_segments to 1 and set the segment boundary to the
same as a page boundary to ensure there is only ever one single-page
segment of IO requested at a time.

This can particularly be an issue when 'xen' is used as it is
known to submit multiple small buffers in a single bio.

Signed-off-by: NeilBrown <neilb@suse.de>
Cc: stable@kernel.org
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>