call md_stop_writes from md_stop
Patchwork-id: 25389
O-Subject: [Patch RHEL6 52/60] md: call md_stop_writes from md_stop
Bugzilla: 583050 586296 586299 588371
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: NeilBrown <neilb@suse.de>

This moves the call to the other side of set_readonly, but that should
not be an issue.
This encapsulates in 'md_stop' all of the functionality for internally
stopping the array, leaving all the interactions with externalities
(sysfs, request_queue, gendisk) in do_md_stop.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>