Include the GNU build ID note in the digest
Patchwork-id: 25411
O-Subject: [RHEL6 PATCH] MODSIGN: Include the GNU build ID note in the digest
Bugzilla: 581965
RH-Acked-by: Roland McGrath <roland@redhat.com>

Include the GNU build ID note in the digest for the kernel module signature [BZ
581965].

This requires that the rpm-build package be updated to 4.8.0-7.el6 on the build
machines before building the kernel RPM otherwise the debugedit program will
change the GNU build ID of the .ko file to include the signature note too,
thereby rendering the signature invalid [BZ 590947].

To this end, the kernel spec file should have a requirement set for this.

This patch must be applied on top of the previously posted patch:

	Subject: [RHEL6 PATCH] MODSIGN: Fix module signing
	Date: Fri, 14 May 2010 16:01:45 +0100

If it is applied, the kernel will only successfully load if debugedit
(rpm-build) was updated on the build machine.  If it wasn't the kernel will
fail with messages about modules failing to load.

It can be checked as to whether this patch was actually applied by:

	echo 1 >/sys/module/kernel/parameters/modsign_debug
	modprobe ext2

where any not-yet-loaded module can be substituted for ext2.  dmesg should be
examined, and a list of sections included in the digest should be seen:

	00000020 00 digested the .bss section, size 16
	00000181 00 digested the .data section, size 320
	000001cf 00 digested the .exit.text section, size 40
	00000434 00 digested the .gnu.linkonce.this_module section, size 560
	00000505 00 digested the .init.text section, size 171
	00000629 00 digested the .modinfo section, size 256
	0000067b 00 digested the .note.gnu.build-id section, size 36
	...

If a section called ".note.gnu.build-id" is listed, then the patch worked.  If
it wasn't, then the patch was not applied or the module does not have a GNU
build ID.

Signed-off-by: David Howells <dhowells@redhat.com>