KVM SVM Implement workaround for Erratum 383
Patchwork-id: 25439
O-Subject: [RHEL6 PATCH 2/2] KVM SVM Implement workaround for Erratum 383
Bugzilla: 592311
RH-Acked-by: Rik van Riel <riel@redhat.com>

This patch is a fix for AMD erratum 383 for family 10h processors.

This erratum affects KVM virtualization, and deals with a malicious guest
trying to gain access by repeatedly trying multiple addresses.  There is a
small window between the MCE generation and the guest gaining access, and
sometimes the timing may be such that the MCE is not generated in time.
The proposed patch will whack such a guest and aborts it.   This prevents a
system data corruption issue.

Resolves BZ 592311

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>