fix race between do_huge_pmd_anonymous_page and pte_alloc_map
Patchwork-id: 25443
O-Subject: [PATCH 51 of 49 RHEL6] Transparent Hugepage Support #20 -> #23 -try2
Bugzilla: 556572
RH-Acked-by: Rik van Riel <riel@redhat.com>

This is to apply on top of the previous.

Just yesterday I found out that by running "cp /dev/vda /dev/null" in
guest with virtio on 6G guest and 4G ram on host, was crashing the
host immediately after the host hit swap, only with
THP+memorycompaction enabled. I just fixed it, it turned out it was a
core THP race that went unnoticed so long. It may have caused other
unexpected issues. I'm now 2G in swap and there are zero problems
after applying this fix. This affects THP as a whole not just KVM and
I've no idea why it wasn't reproducible with IDE (I already tested
this workload with ide and it used to work well and it still worked
fine, but virtio-blk triggers it immediately).

On the good side, the huge_memory.c implementation I think is very
clean and the invariants very clear and not too complex, lots of if
documented in Documentation/transhuge.txt as well and in fact it
wasn't hard to fix this at all.

In any case there was zero risk with THP disabled as I guaranteed
before THP was merged in RHEL6.

Now virtio-blk will work much better with THP enabled...