GFS2: Don't "get" xattrs for ACLs when ACLs are turned off
Patchwork-id: 25467
O-Subject: [RHEL6] GFS2: Don't "get" xattrs for ACLs when ACLs are turned off
	(bz #546294)
Bugzilla: 546294
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

This is to match RHEL5 and ext3 behaviour. I've posted the same
patch to cluster-devel and that will be pushed into the -nmw
tree just as soon as Linus has pulled what is currently there.

This is for bz #546294

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>