gro: Fix bogus gso_size on the first fraglist entry
Patchwork-id: 25501
O-Subject: gro: Fix bogus gso_size on the first fraglist entry
Bugzilla: 594561
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

Hi:

RHEL6 Bugzilla #594561

This patch is for a bug first discovered under RHEL5.  Its effect
under RHEL6 is mitigated by updates to the KVM user-space package.
However, under the right circumstances it can still result in a
kernel crash.

It has been accepted upstream.

gro: Fix bogus gso_size on the first fraglist entry

When GRO produces fraglist entries, and the resulting skb hits
an interface that is incapable of TSO but capable of FRAGLIST,
we end up producing a bogus packet with gso_size non-zero.

This was reported in the field with older versions of KVM that
did not set the TSO bits on tuntap.

This patch fixes that.

Reported-by: Igor Zhang <yugzhang@redhat.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>

Cheers,

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>