Fix FV Xen guest when xen_pv_hvm not enabled
Patchwork-id: 25675
O-Subject: [RHEL6.0 PATCH] Fix FV Xen guest when xen_pv_hvm not enabled
Bugzilla: 523134
RH-Acked-by: Andrew Jones <drjones@redhat.com>

BZ 523134 (original BZ for Xen pv-on-hvm support for rhel6)

Evidently I didn't test a FV Xen guest with pv-on-hvm patches
with the xen_pv_hvm kernel param _not_ set to =enable.

The patch 4/4 of the series introduced this snippet
to xenbus_probe.c:

	-@ -966,6 +983,9 @@ static void wait_for_devices(struct xenbus_driver *xendrv)
	 #ifndef MODULE
	 static int __init boot_wait_for_devices(void)
	 {
	+	if (!xenbus_frontend.error)
	+		return xenbus_frontend.error;
	+
		ready_to_wait_for_devices = 1;
		wait_for_devices(NULL);
		return 0;

which was a cut-n-paste error on my part.

The following patch (against -28 kernel) is the correct snippet,
and dupes what is in rhel5's xenbus_probe.c, for the bugfix
that patch 4/4 was trying to solve:
  Prevent pv drivers from crashing a FV guest if pv-on-hvm not configured.

Built x86_64 only.
Re-tested with following combos:
 (1) Xen FV guest, xen_pv_hvm=enable set on kernel cmdline
 (2) Xen FV guest, xen_pv_hvm _not_ on kernel cmdline
 (3) Xen PV guest, xen_pv_hvm=enable set on kernel cmdline (ensure no regression)
 (4) Xen PV guest, xen_pv_hvm _not_ on kernel cmdline (ensure no regression).

This needs to get into beta 2 build, or else RHEL6 as FV guest on (rhel5) Xen
won't boot.

Please review & ack.

- Don

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>