BUILD: Copy kernel file as hard link and add 'rh-clean' & 'rh-clean-rpmdirs' Makefile targets
Patchwork-id: 25702
O-Subject: [RHEL6 PATCH] BUILD V2: Copy kernel file as hard link and add
	'rh-clean' & 'rh-clean-rpmdirs' Makefile targets
Bugzilla:
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Version 2:
Updated the help text to correctly specify all three rh-clean-* targets
are called for rh-clean.

A patch to have the Makefile copy the linux-2.6.32.tar.bz2 as a
hardlink.  Benefits are that it copies much faster, also when doing an
rsync of the RHEL6 kernel tree to another system using the '-avPH'
options, the rsync transfer will be faster.  If there is an error doing
the link it will fall back to non-link copy command.

I added a 'rh-clean-rpmdirs' target which cleans the
redhat/rpm/{BUILD,SRPMS,RPMS,SPECS}/ directories.

Also added a 'rh-clean' Makefile target.  It has the three rh-clean-*
targets as dependencies, so they will all be called.

Testing:
I have done a 'make rh-clean' & 'make rh-clean-rpmdirs' and it cleaned
the directories as expected.

I did a 'make rh-srpm' and it created a hardlink of linux-2.6.32.tar.bz2
in redhat/rpms/SOURCE/

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>