SUNRPC: Fix a use after free bug with the NFSv4.1 backchannel
Patchwork-id: 25733
O-Subject: [RHEL6 PATCH 1/2] SUNRPC: Fix a use after free bug with the NFSv4.1
	backchannel
Bugzilla: 595478
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

The ->release_request() callback was designed to allow the transport layer
to do housekeeping after the RPC call is done. It cannot be used to free
the request itself, and doing so leads to a use-after-free bug in
xprt_release().

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>