enhance network cgroup classifier to work in softirq context
Patchwork-id: 25738
O-Subject: [RHEL6 PATCH] enhance network cgroup classifier to work in softirq
	context (bz 595351)
Bugzilla: 595351
RH-Acked-by: David S. Miller <davem@redhat.com>

Hey-
	In tinkering with the network cgroup classifier, I found that, since the
classifier relied on the current pointer pointing to the process that sent the
packet being classified, it would not work on any packets after having
transitioned through a softirq context.  Since kvm uses the software bridge and
the tun driver to transmit frames, we are guaranteed at least 1 softirq
transition on that path, meaning that kvm did not work with the network cgroup
controller.  Herbert Xu came up with this solution, in which we tuck the classid
of a frame into its sk structure at the top of the stack, and in the classifier,
if we're in softirq context, we use that field to determine the class of the
frame.  This lets the classifier work in softirq context and with that, kvm.

Fixes bz 595351

Upstream status:
http://git.kernel.org/?p=linux/kernel/git/davem/net-2.6.git;a=commit;h=f845172531fb7410c7fb7780b1a6e51ee6df7d52
http://git.kernel.org/?p=linux/kernel/git/davem/net-2.6.git;a=commit;h=8286274284e15b11b0f531b6ceeef21fbe00a8dd

Brew Status:
https://brewweb.devel.redhat.com/taskinfo?taskID=2467354

Test Status:
I've tested upstream and with the above brew kernel on mark wagners perf system.
Works well.

Neil

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>