Ensure that we mark the inode as dirty if we exit early from commit
Patchwork-id: 25834
O-Subject: [RHEL6 PATCH 3/4] NFS: Ensure that we mark the inode as dirty if we
	exit early from commit
Bugzilla: 595478
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

If we exit from nfs_commit_inode() without ensuring that the COMMIT rpc
call has been completed, we must re-mark the inode as dirty. Otherwise,
future calls to sync_inode() with the WB_SYNC_ALL flag set will fail to
ensure that the data is on the disk.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>