Config update for powerpc
Patchwork-id: 4387
O-Subject: [kernel team] [Patch RHEL6]Config update for powerpc
Bugzilla: 597328

Long ago resolved, or so we thought, issue is that RDS is supposed to be
enabled.  Well, it is, but config-generic-powerpc turns it back off just
on the powerpc arch.  Unfortunately, that's specifically an arch that
IBM cares about.  So, they are saying that bug 463800 hasn't been met
because it's on powerpc that they want RDS enabled.  Doh!  Can't say as
I blame 'em.  Here's the patch to fix this.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>