nfsd4: simplify lease/grace interaction
Patchwork-id: 25850
O-Subject: [RHEL6 PATCH 02/47] nfsd4: simplify lease/grace interaction
Bugzilla: 597215
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: J. Bruce Fields <bfields@citi.umich.edu>

The original code here assumed we'd allow the user to change the lease
any time, but only allow the change to take effect on restart.  Since
then we modified the code to allow setting the lease on when the server
is down.  Update the rest of the code to reflect that fact, clarify
variable names, and add document.

Also, the code insisted that the grace period always be the longer of
the old and new lease periods, but that's overly conservative--as long
as it lasts at least the old lease period, old clients should still know
to recover in time.

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>