sunrpc: don't keep expired entries in the auth caches
Patchwork-id: 25847
O-Subject: [RHEL6 PATCH 07/47] sunrpc: don't keep expired entries in the auth
	caches.
Bugzilla: 597215
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: NeilBrown <neilb@suse.de>

currently expired entries remain in the auth caches as long
as there is a reference.
This was needed long ago when the auth_domain cache used the same
cache infrastructure.  But since that (being a very different sort
of cache) was separated, this test is no longer needed.

So remove the test on refcnt and tidy up the surrounding code.

This allows the cache_dequeue call (which needed to be there to
drop a potentially awkward reference) can be moved outside of the
spinlock which is a better place for it.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>